backend/requirements.txt
Flask>=2.0
Flask-Mail
Flask-SQLAlchemy
Flask-Login
Flask-CORS
PyMySQL
APScheduler
gunicorn
python-dotenv 